+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## v0.2.0
 
 - removed Mantle
+- removed The One Probe
 - added Chicken Drop Feathers
 - added Comforts
 - added FTB Gui Library
 - added Path under gate
 - added Custom Starter Gear
+- added quark oddities
 - lots of item and mechanic changes, that are too many to track here right now. Later version changelogs will track those, when things settled down.